requirements/api.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
xgboost
torch
torch-geometric
shap
sqlalchemy
psycopg[binary]
PyJWT
